Moved HuggingFace test
