+++ tests/test_repo_other.py
@@ -2,6 +2,8 @@
 """Dribdat data aggregation tests."""
 
 from dribdat.api.parser import GetProjectData
+from dribdat.api.parser import get_huggingface_project
+from dribdat.apifetch import FetchHuggingFaceProject
 from requests.exceptions import ReadTimeout
 import warnings
 
@@ -49,3 +51,15 @@ class TestRepoOther:
         assert "commits" in test_obj
         assert len(test_obj["commits"]) > 5
         assert test_obj["commits"][-1]["message"] == "Initial cookiecutter-flask project"
+
+    def test_get_huggingface_project(testapp):
+        """Test getting a Hugging Face project."""
+        url = "https://huggingface.co/google-bert/bert-base-uncased"
+        with testapp.app.app_context():
+            data = get_huggingface_project(url)
+        assert "name" in data
+        assert data["name"] == "google-bert/bert-base-uncased"
+        assert "description" in data
+        assert "Bert" in data["description"]
+        assert "commits" in data
+        assert len(data["commits"]) > 0
